Fix some iOS device no response after login
Dialog is mounted on loading modal, when loading modal dismiss the
dialog is also dismissed, the behaviour is different in different device
especially simulator

Fix:

use simple view to replace loading modal
